% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_ingrain}
\alias{theme_ingrain}
\title{Minimal theme for ingrain figures}
\usage{
theme_ingrain(base_size = 12)
}
\arguments{
\item{base_size}{Base font size in points.}
}
\value{
A ggplot2 theme object.
}
\description{
Minimal theme for ingrain figures
}
\examples{
a <- ingrain(crayfish, grain = 1000)
autoplot(a) + theme_ingrain(base_size = 11)
}
